+++ modules/@apostrophecms/http/ui/apos/big-upload-client.js
@@ -38,7 +38,6 @@ const defaultChunkSize = 1024 * 1024 * 4;
 export default async (url, options) => {
   const chunkSize = options.chunkSize || defaultChunkSize;
   const http = options.http || window.apos?.http;
-  const progress = options.progress || (n => {});
   const files = options.files || {};
   const info = {};
   let totalBytes = 0;
@@ -82,7 +81,9 @@ export default async (url, options) => {
         body: formData
       });
       sentBytes += thisChunkSize;
-      progress(sentBytes / totalBytes);
+      if (typeof options.progress === 'function') {
+        progressInterface(options.progress, sentBytes, totalBytes);
+      }
       chunk++;
     }
     n++;
@@ -99,3 +100,15 @@ export default async (url, options) => {
   });
   return result;
 };
+
+function progressInterface(fn, sent, total) {
+  if (typeof fn !== 'function') {
+    return;
+  }
+  if (fn.length === 1) {
+    fn(sent / total);
+    return;
+  }
+
+  fn(sent, total);
+}